Add support for deep links in Android
We also add a payment request module so we can sync the app launch and the intent processing, making sure the wallet is fully loaded before attempting to open the send screen.

This feature makes it possible to e.g. scan a BIP21 URI from the phone camera qr scanner and open the wallet accordingly.