Bump the toolchain to latest version
Since Nov. 1st the play store requires the apps to be compatible with 16KB memory page sizes. This is only supported in recent enough versions so it's a good excuse to bump everything now.

This implies to switch to the new architecture as well and update the code to comply with the new methods. This still needs some cleanup as several deprecated methods and variables are now used but this is out of scope for this large enough commit.

I also added an android clean rule for convenience.